Separate warning and status checks
Replace the shared multi-mode checker with one script for Learn warning comparison and one script for the automerge all-green gate.
